// projects/ui-kit/ui-components/button/src/types.ts
export type ButtonTypes = 'neutral' | 'outline' | 'ghost' | 'link';
export type ButtonShapes = 'neutral' | 'circle' | 'square';
export type ButtonSizes =
  | 'neutral'
  | 'xs'
  | 'sm'
  | 'md'
  | 'lg'
  | 'wide'
  | 'block';
export type ButtonColors =
  | 'neutral'
  | 'primary'
  | 'secondary'
  | 'accent'
  | 'success'
  | 'warning'
  | 'error'
  | 'info';